+++ gui/MainWindowCentralWidget.py
@@ -14,6 +14,9 @@ class MainWindowCentralWidget(QWidget):
     update_save_button_requested = pyqtSignal(bool)
 
     def __init__(self, parent=None):
+        """Central widget for the main window. Handles the open tabs.
+            :param parent: The parent widget (actually an IDEMainWindow) to attach this widget to.
+        """
         super().__init__(parent)
 
         self.tab_widget = QTabWidget()
@@ -32,10 +35,13 @@ class MainWindowCentralWidget(QWidget):
 
         self.pywright_installation_path: str = ""
         self.selected_game_info: PyWrightGameInfo | None = None
+        self._game_properties_widget: GamePropertiesWidget | None = None
 
     def set_selected_game(self, selected_game_info: PyWrightGameInfo):
         self.selected_game_info = selected_game_info
         self.pywright_installation_path = str(self.selected_game_info.pywright_folder_path)
+        self._game_properties_widget = GamePropertiesWidget(self.selected_game_info)
+        self._game_properties_widget.load_game(self.selected_game_info)
 
     # ====== Tab Handling ======
     # ==== Opening new tab ====
@@ -44,20 +50,20 @@ class MainWindowCentralWidget(QWidget):
         self.tab_widget.setMovable(self.tab_widget.count() > 1)
         self.tab_widget.setCurrentIndex(self.tab_widget.count() - 1)
 
-    def open_game_properties_tab(self, game_properties_widget: GamePropertiesWidget):
+    def open_game_properties_tab(self):
         for i in range(self.tab_widget.count()):
-            if self._is_game_properties_tab(i):
+            if self.is_game_properties_tab(i):
                 # We already have a Game Properties tab open, so switch to that instead.
                 self.tab_widget.setCurrentIndex(i)
                 return
 
-        self.open_new_tab(game_properties_widget, "Game Properties")
+        self.open_new_tab(self._game_properties_widget, "Game Properties")
 
     def open_new_editing_tab(self, file_path: str):
         # Don't open a file editing tab if there is already one open for a given file
         for i in range(self.tab_widget.count()):
             # Skip Game Properties tab
-            if self._is_game_properties_tab(i):
+            if self.is_game_properties_tab(i):
                 continue
 
             opened_tab: FileEditWidget = self.tab_widget.widget(i)
@@ -84,12 +90,12 @@ class MainWindowCentralWidget(QWidget):
     # ==== Changing to a new tab ====
     def _handle_tab_change(self, index: int):
         condition = self.tab_widget.count() > 0 and \
-                    not self._is_game_properties_tab(index) and \
+                    not self.is_game_properties_tab(index) and \
                     self.tab_widget.currentWidget().is_file_modified()
 
         self.update_save_button_requested.emit(condition)
 
-    def _is_game_properties_tab(self, index: int):
+    def is_game_properties_tab(self, index: int):
         return isinstance(self.tab_widget.widget(index), GamePropertiesWidget)
 
     # ==== Renaming a tab ====
@@ -98,7 +104,7 @@ class MainWindowCentralWidget(QWidget):
 
     # ==== Saving a tab ====
     def handle_save_tab(self):
-        if not self._is_game_properties_tab(self.tab_widget.currentIndex()):
+        if not self.is_game_properties_tab(self.tab_widget.currentIndex()):
             self.tab_widget.currentWidget().save_to_file()
             self._update_save_button_and_current_tab()
 
@@ -108,7 +114,7 @@ class MainWindowCentralWidget(QWidget):
     def _update_tab_modified_infos(self):
         # Like _update_save_button() but for all tabs
         for idx in range(self.tab_widget.count()):
-            if self._is_game_properties_tab(idx):
+            if self.is_game_properties_tab(idx):
                 continue
 
             self._update_file_editing_tab_info(idx)
@@ -124,7 +130,7 @@ class MainWindowCentralWidget(QWidget):
 
     # ==== Removing a tab ====
     def _handle_remove_tab(self, index):
-        if not self._is_game_properties_tab(index):
+        if not self.is_game_properties_tab(index):
             tab: FileEditWidget = self.tab_widget.widget(index)
 
             if tab.is_file_modified():
@@ -144,7 +150,7 @@ class MainWindowCentralWidget(QWidget):
         result = []
 
         for tab_idx in range(self.tabs_count()):
-            if self._is_game_properties_tab(tab_idx):
+            if self.is_game_properties_tab(tab_idx):
                 result.append("Game Properties")
             else:
                 file_widget: FileEditWidget = self.tab_widget.widget(tab_idx)
@@ -183,7 +189,7 @@ class MainWindowCentralWidget(QWidget):
             # If nothing is open, inform the user and do nothing.
             QMessageBox.information(self, "Find/Replace", "There are no tabs open.")
             return
-        if not self._is_game_properties_tab(self.tab_widget.currentIndex()):
+        if not self.is_game_properties_tab(self.tab_widget.currentIndex()):
             file_widget: FileEditWidget = self.tab_widget.currentWidget()
             file_widget.search_in_file(text, find_type, search_scope)
 
@@ -196,7 +202,7 @@ class MainWindowCentralWidget(QWidget):
         if replace_type == ReplaceType.REPLACE_ALL and search_scope == SearchScope.OPEN_TABS:
             self.replace_all_in_all_open_tabs(text_to_find, text_to_replace)
             return
-        if not self._is_game_properties_tab(self.tab_widget.currentIndex()):
+        if not self.is_game_properties_tab(self.tab_widget.currentIndex()):
             file_widget: FileEditWidget = self.tab_widget.currentWidget()
             file_widget.replace_in_file(text_to_find, text_to_replace, replace_type, search_scope)
 
@@ -215,7 +221,7 @@ class MainWindowCentralWidget(QWidget):
             return
 
         for idx in range(current_position + 1, tabs_count):
-            if self._is_game_properties_tab(idx):
+            if self.is_game_properties_tab(idx):
                 continue
 
             curr_widget: FileEditWidget = self.tab_widget.widget(idx)
@@ -243,7 +249,7 @@ class MainWindowCentralWidget(QWidget):
             return
 
         for idx in range(current_position).__reversed__():
-            if self._is_game_properties_tab(idx):
+            if self.is_game_properties_tab(idx):
                 continue
 
             curr_widget: FileEditWidget = self.tab_widget.widget(idx)
@@ -273,7 +279,7 @@ class MainWindowCentralWidget(QWidget):
             return
 
         for idx in range(current_position + 1, tabs_count):
-            if self._is_game_properties_tab(idx):
+            if self.is_game_properties_tab(idx):
                 continue
 
             curr_widget: FileEditWidget = self.tab_widget.widget(idx)
@@ -292,7 +298,7 @@ class MainWindowCentralWidget(QWidget):
         tab_count = self.tab_widget.count()
 
         for idx in range(tab_count):
-            if self._is_game_properties_tab(idx):
+            if self.is_game_properties_tab(idx):
                 continue
 
             curr_tab: FileEditWidget = self.tab_widget.widget(idx)
@@ -305,10 +311,35 @@ class MainWindowCentralWidget(QWidget):
     def clear_tabs(self):
         self.tab_widget.clear()
 
+    def restore_last_open_tabs(self, open_tabs: list[str], last_tab_index: int) -> list[str]:
+        """Tries to restore last open tabs if there are any, and returns a list of missing files.
+            :param open_tabs: List of previously opened tabs acquired from IDE settings
+            :param last_tab_index: Last open tab's index acquired from IDE settings
+            :return: A list of missing files (or an empty list if there are none)"""
+        missing_tabs = []
+
+        for tab_path in open_tabs:
+            if tab_path == "Game Properties":
+                self.open_game_properties_tab()
+            elif Path(tab_path).exists() and Path(tab_path).is_file():
+                self.open_new_editing_tab(tab_path)
+            else:
+                # File on tab_path is missing.
+                missing_tabs.append(tab_path)
+                last_tab_index -= 1
+
+        if last_tab_index < 0:
+            last_tab_index = 0
+
+        self.tab_widget.setCurrentIndex(last_tab_index)
+
+        # Emit files missing signal if there are any files missing
+        return missing_tabs
+
     def apply_settings(self):
         EditorThemes.current_editor_theme.load_theme(IDESettings.get_editor_color_theme())
         for idx in range(self.tabs_count()):
-            if not self._is_game_properties_tab(idx):
+            if not self.is_game_properties_tab(idx):
                 tab: FileEditWidget = self.tab_widget.widget(idx)
                 tab.supply_font_properties_to_lexer(IDESettings.get_font_name(),
                                                     IDESettings.get_font_size(),
@@ -323,7 +354,7 @@ class MainWindowCentralWidget(QWidget):
             return
 
         # Don't do anything if Game Properties is open
-        if self._is_game_properties_tab(self.tab_widget.currentIndex()):
+        if self.is_game_properties_tab(self.tab_widget.currentIndex()):
             QMessageBox.critical(self, "Error", "Current tab is not a script editing tab!")
             return
 
@@ -335,7 +366,7 @@ class MainWindowCentralWidget(QWidget):
         result = []
 
         for i in range(self.tabs_count()):
-            if self._is_game_properties_tab(i):
+            if self.is_game_properties_tab(i):
                 continue
 
             tab: FileEditWidget = self.tab_widget.widget(i)