+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Fix the crash caused by tab(s) being unable to be restored due to file(s) no longer existing (by in1tiate)
+* PyWright IDE now informs the user in the case of some tabs being unable to be restored due to missing files.
 
 ## Version 1.4 - 06.05.2025